# docs/swagger.yaml
openapi: 3.0.3
info:
  title: TeamSupportPro API
  description: REST API for TeamSupportPro — a team-based IT support ticket management system.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: team_support_session
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ActivityEntry:
      type: object
      properties:
        id: { type: string }
        ticketId: { type: string }
        actor: { type: string }
        message: { type: string }
        at: { type: string, format: date-time }
    Ticket:
      type: object
      properties:
        id: { type: string, example: "TKT-10001" }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [Open, In Progress, Pending, Resolved, Closed] }
        priority: { type: string, enum: [Low, Medium, High, Critical] }
        teamId: { type: string }
        categoryId: { type: string }
        assignedToId: { type: string, nullable: true }
        requestorName: { type: string }
        requestorEmail: { type: string }
        location: { type: string }
        dueLabel: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        activity:
          type: array
          items: { $ref: '#/components/schemas/ActivityEntry' }
...
    Error:
      type: object
      properties:
        error: { type: string }
    Location:
      type: object
      properties:
        id: { type: string, example: "loc-a1b2c3" }
        name: { type: string, example: "Adams Elementary School" }
        isActive: { type: boolean, example: true }
        sortOrder: { type: integer, example: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
security:
  - cookieAuth: []
  - apiKeyHeader: []
  - bearerAuth: []
paths:
...
  /api/settings/locations:
    get:
      summary: List all locations (admin)
      description: Returns all locations including inactive ones. Requires admin authentication.
      responses:
        "200":
          description: All locations
          content:
            application/json:
              schema:
                type: object
                properties:
                  locations:
                    type: array
                    items: { $ref: '#/components/schemas/Location' }
    post:
      summary: Create a location (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "Adams Elementary School" }
                sortOrder: { type: integer, example: 0 }
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                type: object
                properties:
                  location: { $ref: '#/components/schemas/Location' }
        "409":
          description: Name conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/tickets:
    get:
      summary: List tickets for the authenticated user's team
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items: { $ref: '#/components/schemas/Ticket' }
        "401":
          description: Unauthenticated
    post:
      summary: Create a new ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, priority, teamId, categoryId, requestorName, requestorEmail]
              properties:
                id: { type: string }
                title: { type: string }
                description: { type: string }
                priority: { type: string, enum: [Low, Medium, High, Critical] }
                teamId: { type: string }
                categoryId: { type: string }
                assignedToId: { type: string, nullable: true }
                requestorName: { type: string }
                requestorEmail: { type: string }
                location: { type: string }
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket: { $ref: '#/components/schemas/Ticket' }
        "400":
          description: Invalid input
        "401":
          description: Unauthenticated
        "403":
          description: Cross-team ticket creation forbidden
...
  /api/public/tickets:
    post:
      summary: Create an anonymous/public ticket
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId, categoryId, title, description, requestorName, requestorEmail]
              properties:
                teamId: { type: string }
                categoryId: { type: string }
                title: { type: string }
                description: { type: string }
                requestorName: { type: string }
                requestorEmail: { type: string }
                location: { type: string }
      responses:
        "201":
          description: Public ticket created
        "400":
          description: Invalid input
...